Backend/app/requirements.txt
Flask
mysql-connector
requests
pandas
numpy